rpc: resolve implicit type conversion warning
(cherry picked from commit 58c11066c4898090c16d3b3b576f48b5ef016085)

+++ main/utils/cbor_rpc.c
@@ -552,7 +552,7 @@ bool rpc_get_bip32_path_from_value(CborValue* value, uint32_t* path_ptr, const s
         if (tmp > 0xFFFFFFFF) {
             return false;
         }
-        path_ptr[counter] = tmp;
+        path_ptr[counter] = (uint32_t)tmp;
 
         cberr = cbor_value_advance_fixed(&arrayItem);
         JADE_ASSERT(cberr == CborNoError);
